+++ typescript/core_api_bindings/api_container_service_grpc_pb.js
@@ -38,6 +38,28 @@ function deserialize_api_container_api_ExecuteBulkCommandsArgs(buffer_arg) {
   return api_container_service_pb.ExecuteBulkCommandsArgs.deserializeBinary(new Uint8Array(buffer_arg));
 }
 
+function serialize_api_container_api_ExecuteLambdaArgs(arg) {
+  if (!(arg instanceof api_container_service_pb.ExecuteLambdaArgs)) {
+    throw new Error('Expected argument of type api_container_api.ExecuteLambdaArgs');
+  }
+  return Buffer.from(arg.serializeBinary());
+}
+
+function deserialize_api_container_api_ExecuteLambdaArgs(buffer_arg) {
+  return api_container_service_pb.ExecuteLambdaArgs.deserializeBinary(new Uint8Array(buffer_arg));
+}
+
+function serialize_api_container_api_ExecuteLambdaResponse(arg) {
+  if (!(arg instanceof api_container_service_pb.ExecuteLambdaResponse)) {
+    throw new Error('Expected argument of type api_container_api.ExecuteLambdaResponse');
+  }
+  return Buffer.from(arg.serializeBinary());
+}
+
+function deserialize_api_container_api_ExecuteLambdaResponse(buffer_arg) {
+  return api_container_service_pb.ExecuteLambdaResponse.deserializeBinary(new Uint8Array(buffer_arg));
+}
+
 function serialize_api_container_api_GenerateFilesArgs(arg) {
   if (!(arg instanceof api_container_service_pb.GenerateFilesArgs)) {
     throw new Error('Expected argument of type api_container_api.GenerateFilesArgs');
@@ -80,6 +102,28 @@ function serialize_api_container_api_GetServiceInfoResponse(arg) {
 
 function deserialize_api_container_api_GetServiceInfoResponse(buffer_arg) {
   return api_container_service_pb.GetServiceInfoResponse.deserializeBinary(new Uint8Array(buffer_arg));
+}
+
+function serialize_api_container_api_LoadModuleArgs(arg) {
+  if (!(arg instanceof api_container_service_pb.LoadModuleArgs)) {
+    throw new Error('Expected argument of type api_container_api.LoadModuleArgs');
+  }
+  return Buffer.from(arg.serializeBinary());
+}
+
+function deserialize_api_container_api_LoadModuleArgs(buffer_arg) {
+  return api_container_service_pb.LoadModuleArgs.deserializeBinary(new Uint8Array(buffer_arg));
+}
+
+function serialize_api_container_api_LoadModuleResponse(arg) {
+  if (!(arg instanceof api_container_service_pb.LoadModuleResponse)) {
+    throw new Error('Expected argument of type api_container_api.LoadModuleResponse');
+  }
+  return Buffer.from(arg.serializeBinary());
+}
+
+function deserialize_api_container_api_LoadModuleResponse(buffer_arg) {
+  return api_container_service_pb.LoadModuleResponse.deserializeBinary(new Uint8Array(buffer_arg));
 }
 
 function serialize_api_container_api_LoadStaticFilesArgs(arg) {
@@ -194,6 +238,30 @@ function deserialize_google_protobuf_Empty(buffer_arg) {
 
 
 var ApiContainerServiceService = exports.ApiContainerServiceService = {
+  // Starts a module container into the network
+loadModule: {
+    path: '/api_container_api.ApiContainerService/LoadModule',
+    requestStream: false,
+    responseStream: false,
+    requestType: api_container_service_pb.LoadModuleArgs,
+    responseType: api_container_service_pb.LoadModuleResponse,
+    requestSerialize: serialize_api_container_api_LoadModuleArgs,
+    requestDeserialize: deserialize_api_container_api_LoadModuleArgs,
+    responseSerialize: serialize_api_container_api_LoadModuleResponse,
+    responseDeserialize: deserialize_api_container_api_LoadModuleResponse,
+  },
+  // Executes a Kurtosis Lambda module function on behalf of the user 
+executeLambda: {
+    path: '/api_container_api.ApiContainerService/ExecuteLambda',
+    requestStream: false,
+    responseStream: false,
+    requestType: api_container_service_pb.ExecuteLambdaArgs,
+    responseType: api_container_service_pb.ExecuteLambdaResponse,
+    requestSerialize: serialize_api_container_api_ExecuteLambdaArgs,
+    requestDeserialize: deserialize_api_container_api_ExecuteLambdaArgs,
+    responseSerialize: serialize_api_container_api_ExecuteLambdaResponse,
+    responseDeserialize: deserialize_api_container_api_ExecuteLambdaResponse,
+  },
   // Registers a service with the API container but doesn't start the container for it
 registerService: {
     path: '/api_container_api.ApiContainerService/RegisterService',